+++ packages/node/src/configure/SubqueryProject.ts
@@ -175,7 +175,7 @@ async function loadProjectFromManifest1_0_0(
 }
 
 async function loadProjectTemplates(
-  projectManifest: ProjectManifestV1_0_0Impl | ProjectManifestV1_0_0Impl,
+  projectManifest: ProjectManifestV1_0_0Impl,
   root: string,
   reader: Reader,
 ): Promise<SubqlProjectDsTemplate[]> {