+++ packages/node/src/indexer/dynamic-ds.service.ts
@@ -53,7 +53,7 @@ export class DynamicDsService {
 
       return ds;
     } catch (e) {
-      logger.error(e.message);
+      logger.error(e, 'Failed to create dynamic ds');
       process.exit(1);
     }
   }
@@ -67,7 +67,7 @@ export class DynamicDsService {
           params.map((params) => this.getDatasource(params)),
         );
       } catch (e) {
-        logger.error(`Unable to get dynamic datasources:\n${e.message}`);
+        logger.error(e, `Unable to get dynamic datasources`);
         process.exit(1);
       }
     }