+++ packages/node/src/yargs.ts
@@ -151,6 +151,18 @@ export function getYargsOption() {
       describe: 'Disable storing historical state entities',
       type: 'boolean',
     },
+    reindex: {
+      demandOption: false,
+      describe: 'Reindex to specified block height',
+      type: 'number',
+    },
+    workers: {
+      alias: 'w',
+      demandOption: false,
+      describe:
+        'Number of worker threads to use for fetching and processing blocks. Disabled by default.',
+      type: 'number',
+    },
   });
 }